빌드 테스트

+++ Roguelike-game-v2/Assets/Scripts/UI/ExpSlider_Main_UI.cs
@@ -16,7 +16,7 @@ public class ExpSlider_Main_UI : UserInterface
     {
         if(Managers.UserData.data.Level != UserLevelInfo_SO.maxLevel)
         {
-            expSlider.value = Managers.UserData.data.Exp / Managers.UserData.UserLevelInfo.requiredEXP[Managers.UserData.data.Level - 1];
+            expSlider.value = (float)Managers.UserData.data.Exp / Managers.UserData.UserLevelInfo.requiredEXP[Managers.UserData.data.Level - 1];
             expText.text = $"{Managers.UserData.data.Exp:N0} / {Managers.UserData.UserLevelInfo.requiredEXP[Managers.UserData.data.Level - 1]:N0}";
         }
         else
